setup part2

+++ risorse/appunti/script/setup.sh
@@ -5,8 +5,11 @@ set -e
 SETUP_FILE=/var/www/html/risorse/files/CONFIGURAZIONI/global_env.sh
 MV_PATH=/opt/script/
 # Verifica se il file è presente
-if [ -f "${SETUP_FILE}" ]; then
+if [ -f "${SETUP_FILE}" ] && [ $(find "${SETUP_FILE}" -mmin -5 | wc -l) -gt 0 ]; then
     mv ${SETUP_FILE} ${MV_PATH}
+
+    # Configura client duckdns update
+    ./setup_duckdns.sh
 
     # crea domini certs
     ./setup_letsencrypt.sh
@@ -18,22 +21,19 @@ if [ -f "${SETUP_FILE}" ]; then
     ./setup_replace_nginx.sh
 
     #avvia nginx
-    systemctl start nginx
-    systemctl enable nginx
+    systemctl restart nginx
 
     # applica trasformazione tomcat
     ./setup_replace_tomcat.sh
 
     #avvia tomcat10
-    systemctl start tomcat10
-    systemctl enable tomcat10
+    systemctl restart tomcat10
 
     # applica trasformazione tomcat
     ./setup_replace_apache.sh
 
     #avvia tomcat10
-    systemctl start apache2
-    systemctl enable apache2
+    systemctl restart apache2
 
     echo "Installazione Completata";
